full change of merging models.
