+++ python-backend/app/api/documents/expected/resources/documents.py
@@ -5,7 +5,7 @@ from datetime import datetime
 from flask import request
 from flask_restplus import Resource, reqparse
 
-from ..models.document import ExpectedDocument
+from ..models.mine_expected_document import MineExpectedDocument
 
 from app.extensions import api
 from ....utils.access_decorators import requires_role_mine_view, requires_role_mine_create, requires_any_of, MINE_VIEW, MINE_CREATE, MINESPACE_PROPONENT
@@ -26,7 +26,7 @@ class ExpectedDocumentResource(Resource, UserMixin, ErrorMixin):
     def get(self, exp_doc_guid=None):
         if exp_doc_guid is None:
             return self.create_error_payload(404, 'Must provide a expected document guid.'), 404
-        mine_exp_doc = ExpectedDocument.find_by_exp_document_guid(exp_doc_guid)
+        mine_exp_doc = MineExpectedDocument.find_by_exp_document_guid(exp_doc_guid)
         if mine_exp_doc is None:
             return self.create_error_payload(404, 'Expected document not found'), 404
         return {'expected_document': mine_exp_doc.json()}
@@ -37,16 +37,15 @@ class ExpectedDocumentResource(Resource, UserMixin, ErrorMixin):
         if exp_doc_guid is None:
             return self.create_error_payload(404, 'Must provide a expected document guid.'), 404
 
-        exp_doc = ExpectedDocument.find_by_exp_document_guid(exp_doc_guid)
+        exp_doc = MineExpectedDocument.find_by_exp_document_guid(exp_doc_guid)
         if exp_doc is None:
             return self.create_error_payload(
-            404, f'expected_document with guid "{exp_doc_guid}" not found'), 404
-        
+                404, f'expected_document with guid "{exp_doc_guid}" not found'), 404
+
         data = self.parser.parse_args()
         updated_doc = data['document']
         if str(exp_doc.exp_document_guid) != updated_doc['exp_document_guid']:
-            return self.create_error_payload(500,
-                                                'exp_document does not match guid provided'), 500
+            return self.create_error_payload(500, 'exp_document does not match guid provided'), 500
 
         exp_doc.exp_document_name = updated_doc.get('exp_document_name')
         exp_doc.exp_document_description = updated_doc.get('exp_document_description')
@@ -65,13 +64,12 @@ class ExpectedDocumentResource(Resource, UserMixin, ErrorMixin):
         exp_doc.save()
         return {'expected_document': exp_doc.json()}
 
-
     @api.doc(params={'exp_doc_guid': 'Required: Mine number or guid. Deletes expected document.'})
     @requires_role_mine_create
     def delete(self, exp_doc_guid=None):
         if exp_doc_guid is None:
             return self.create_error_payload(404, 'Must provide a expected document guid.'), 404
-        exp_doc = ExpectedDocument.find_by_exp_document_guid(exp_doc_guid)
+        exp_doc = MineExpectedDocument.find_by_exp_document_guid(exp_doc_guid)
         if exp_doc is not None:
             exp_doc.active_ind = False
             exp_doc.save()
